NavBar, ViewUserReservations
Added new page for users to view their reservations.

Updated navbar to only show for admin/mangers and to be static at the top of the screen.

Updated security config to allow rooms/search and roomtypes to be accessed if signed in.

Header is now sticky

Updated Dashboard views to show more correct information than it was before

LOTS of polish still required.